initial commit
relates #3 and relates #4 I've created the basic boilerplate code and some CSS fundamentals. The previous project started to feel a bit overcrowded with the fundamentals so I tried separating those into a different file on this project, I'm hopping that will keep it easy to read. In the end I also added a variable with the fetch call to the Teleport API stored in it. I left the Police API untouched so you can have a go at this as well before we start messing with DOM through the requests. The README has been updated in a different commit but I will keep adding stuff as we go and you can of course do the same.