added function to fetch details from a search result
relates #8 the new function, fetchCityDetails, allows us to fetch the details about any of the cities that matched the users initial search. From here we can display specifics onto the page and also search the police api with the latlon key in the city object. I also added a button to the printSearchResults function and made other stylistic changes to keep the page layout readable as we test results