resolving issues from code review
relates #30 got rid of extra scrolling on the landing page and reviewed the use of printHTML(). In some cases using that function would require a restructuring of the code overall that, considering the time constraint and scope of the project, is not worth going through, so I left it as is on those. The loading feedback was added in the last merge and there is some error handling being added in a different commit from Carl that should then close the issue completely.